21.9.2024 Norminette.

+++ libft/src_libft/ft_bzero.c
@@ -18,7 +18,7 @@
 
 #include "../include/libft.h"
 
-void	ft_bzero(void *s, int n)
+void	ft_bzero(void *s, size_t n)
 {
 	char	*temp;
