+++ libft/src_libft/ft_memchr.c
@@ -23,7 +23,7 @@ RETURN VALUE
 
 #include "../include/libft.h"
 
-void	*ft_memchr(const void *s, int c, int n)
+void	*ft_memchr(const void *s, int c, size_t n)
 {
 	const unsigned char	*src;
 	unsigned char		cc;
