+++ libft/src_libft/ft_memset.c
@@ -21,7 +21,7 @@ RETURN VALUE
 
 #include "../include/libft.h"
 
-void	*ft_memset(void *str, int c, int n)
+void	*ft_memset(void *str, int c, size_t n)
 {
 	char	*temp;
 	int		i;
