+++ libft/src_libft/ft_strtrim.c
@@ -44,8 +44,8 @@ char	*ft_strtrim(char const *s1, char const *set)
 {
 	char		*trim;
 	char		*t_trim;
-	int		start;
-	int		end;
+	size_t		start;
+	size_t		end;
 
 	if (!s1 || !set)
 		return (NULL);
